+++ CHANGELOG.md
@@ -5,7 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-### [2.0.0rc1] - [ADD DATE]
+### Unreleased Changes
+
+...
+
+### [2.0.0] - 2018-08-16
+
+#### Added
+
+ * Loading remote inventories from the CLI now will
+   perform an automatic walk along the directory structure of the
+   provided URL, searching for the objects.inv file of the
+   documentation set.
+ * The above URL walking functionality is exposed for API use
+   at sphobjinv.fileops.urlwalk.
+
+#### Changed
+
+ * Instances of DataObjStr and DataObjBytes are NO LONGER IMMUTABLE.
+   Working with Inventory instances was going to be far too cumbersome
+   with immutable DataObjStr instances in .objects.
+
+
+### [2.0.0rc1] - 2018-05-18
 
 #### Fixed